Refactor main.dart to use modular app start
Move the main application logic into a separate file (`app_start/app_start.dart`)
to improve modularity and maintainability. This change includes removing the
previous `MyApp` and `MyHomePage` classes from `main.dart` and importing the
new `AppStart` class for initializing the app. This refactor aims to simplify
the entry point of the application and prepare the codebase for future
extensions or modifications.