Add image asset constant for app logo
Created a new file `app_images.dart` under the `lib/utils/constant` directory to define a constant for the app logo image asset. This constant, named `imgLogo`, points to the path `assets/img/logo.png`. This addition standardizes the way image assets are referenced throughout the application, making it easier to manage and update image paths in the future.