Add authentication wrapper to handle user session state
This commit introduces a new file, `wrapper.dart`, within the `lib/view/auth` directory. The `Wrapper` class is a stateless widget designed to manage the authentication state of the user. It leverages the `Provider` package to access the authentication controller (`ControllerAuth`) and uses a `StreamBuilder` to monitor changes in the authentication state.

When the connection state is waiting, it displays a loading indicator (`AppLoading` with `TypeLoading.page`). Once the authentication state is determined, it routes the user to the `HomeView` if authenticated, or to the `SingInView` if not. This enhancement centralizes authentication handling, improving the app's navigation and user experience.