Add password input field widget for authentication
This commit introduces a new widget, `AuthFieldPass`, designed for handling password input during the authentication process in a Flutter application. The widget is part of the authentication view and utilizes various utilities and controllers to manage password validation and display.

The `AuthFieldPass` widget is a `StatelessWidget` that can be configured to either request a regular password or confirm a password by setting the `isConfirm` flag. It leverages the `provider` package to access the `ControllerAuth` for state management and updates. The widget includes a custom text form field with specific properties for password input, such as a label, icons for visibility toggling, and validation logic.

Validation is handled by `AppValidators`, which checks either the password strength or confirmation match against the current password set in the `ControllerAuth`. The widget also allows for the password to be set directly into the user authentication model upon save.

This addition enhances the authentication flow by providing a dedicated and configurable widget for password input, improving code organization and reusability.