Add SectionHeader widget for home view
This commit introduces a new `SectionHeader` widget to the home view section of the application. The widget is designed to display a section title and optionally includes a button for additional actions. It is implemented as a stateless widget, accepting a required `title` and an optional `onPressed` callback function. The widget is structured to be flexible and can be easily integrated into various sections of the home view.